requirements.txt
pandas
numpy
pydantic
pyyaml
click
pytest
coverage
ruff
black
pre-commit
great-expectations
